requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.7
pydantic>=1.10.7
passlib>=1.7.4
python-jose>=3.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
bcrypt>=4.0.1
alembic>=1.10.2